requirements.txt
amqp==1.4.8
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.19
Django==1.9
django-celery==3.1.17
django-crispy-forms==1.5.2
django-model-utils==2.4
image==1.4.3
kombu==3.0.32
Pillow==3.0.0
pytz==2015.7
redis==2.10.5
